Create bin/mock-sensor for simulating a track sensor device and sending fake times
Each TrackSensor class knows how to create fake/example time data for itself now.

Updated the TrackSensor::MicroWizardFastTrack::TIMES_REGEX to handle
potential \r characters that may be sent by the device.

The device created by mock-sensor is is not in the device_glob and
therefore does not get detected for use with the app yet.